+++ src/components/domain/index.ts
@@ -1,2 +1,4 @@
 export { default as MainHeader } from './MainHeader';
 export { default as Map } from './Map';
+export { default as OnLiveFollowings } from './OnLiveFollowings/OnLiveFollowings';
+export { default as OnLiveFollowingsAtStore } from './OnLiveFollowingsAtStore/OnLiveFollowingsAtStore';
